+++ src/Tests/Given_controller_accepting_object.cs
@@ -1,14 +1,32 @@
+using System;
 using System.IO;
 using System.Linq;
 using Isop;
 using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
-using Tests.FakeControllers;
-namespace Tests.ArgumentParsers
+
+namespace Tests
 {
     [TestFixture]
     public class Given_controller_accepting_object
     {
+        class MyObjectController
+        {
+            public class Argument
+            {
+                public string param1 { get; set; }
+                public string param2 { get; set; }
+                public int param3 { get; set; }
+                public decimal param4 { get; set; }
+            }
+
+            public MyObjectController()
+            {
+                OnAction = (a) => string.Empty;
+            }
+            public Func<Argument, string> OnAction { get; set; }
+            public string Action(Argument a) { return OnAction(a); }
+        }
         [Test,Ignore("Brown")]
         public void It_can_parse_class_and_method_with_object_and_execute()
         {
